refactor: renomear campo 'qrcode_or_pairingcode' para 'qrcode' e ajustar lógica de criação de instâncias

+++ src/services/createInstances.js
@@ -4,7 +4,8 @@ const {
   default: makeWASocket,
   DisconnectReason,
   useMultiFileAuthState,
-  fetchLatestBaileysVersion } = require('@whiskeysockets/baileys');
+  fetchLatestBaileysVersion
+} = require('@whiskeysockets/baileys');
 const QRCode = require('qrcode');
 const { S3Client, PutObjectCommand } = require('@aws-sdk/client-s3');
 const fs = require('fs-extra');
@@ -20,7 +21,7 @@ const S3 = new S3Client({
   },
 });
 
-const createInstances = async (instance, phoneNumber = null) => {
+const createInstances = async (instance) => {
   const { version, isLatest } = await fetchLatestBaileysVersion();
   console.log(`Usando versão do WhatsApp Web: ${version}, atual? ${isLatest}`);
 
@@ -33,82 +34,110 @@ const createInstances = async (instance, phoneNumber = null) => {
   });
 
   sock.ev.on('creds.update', saveCreds);
-  sock.ev.on('connection.update', async (update) => {
-    const { connection, qr } = update;
 
-    if (qr && !phoneNumber) {
-      await handleQr(update, instance);
-    }
-
-    if (connection === 'open') {
-      await handleOpen(instance);
-    }
-
-    if (connection === 'close') {
-      await handleClose(update, instance, phoneNumber);
-    }
-
-    if (connection === 'connecting' && phoneNumber && !sock.authState.creds?.registered) {
-      console.log('Estado do sock.authState.creds:', sock.authState.creds);
-      await handlePairing(instance, phoneNumber);
+  // ✅ Espera o evento de sincronização 100%
+  sock.ev.on('messaging-history.set', async ({ progress }) => {
+    if (progress === 100) {
+      syncCompleted = true;
+      console.log('✅ Sincronização de mensagens concluída (progress = 100)');
+      await finalizeInstance(instance); // 🔄 chama função que envia p/ S3 e encerra sessão
     }
   });
 
-  return { message: `Instância ${instance} criada com sucesso!`, instance, sock };
+  sock.ev.on('connection.update', async (update) => {
+    const { connection, qr } = update;
+
+    if (qr) await handleQr(update, instance);
+
+    if (connection === 'open') await handleOpen(instance);
+    if (connection === 'close') await handleClose(update, instance);
+  });
+
+  return { 
+    message: `Instância criada com sucesso!`, 
+    instance: instance,
+  };
 };
 
 const handleQr = async (update, instance) => {
   const { qr } = update;
   const dataQrCode = await QRCode.toDataURL(qr);
   const expires = new Date(Date.now() + 5 * 60 * 1000);
+
   const payload = {
     status: 'QrCode',
-    qrcode_or_pairingcode: dataQrCode,
+    qrcode: dataQrCode,
     qr_expires_at: expires,
     updated_at: connec.fn.now(),
   };
 
-  const exists = await connec('instances').where({ instance }).first();
-  if (exists) await connec('instances').where({ instance }).update(payload);
-  else await connec('instances').insert({ instance, ...payload });
+  const exists = await connec('instances')
+    .where({ instance })
+    .first();
+  if (exists) {
+    await connec('instances')
+      .where({ instance })
+      .update(payload);
+  } else {
+    await connec('instances')
+      .insert({ instance, ...payload });
+  }
+};
+
+const getQrCode = async (instance) => {
+  const Qrcode = await connec('instances')
+    .where({ instance })
+    .select('qrcode')
+    .first();
+
+  return Qrcode;
+};
+
+const timeForQrCode = async (instance, maxRetries = 10, delay = 500) => {
+  for (let i = 0; i < maxRetries; i++) {
+    const qrcode = await getQrCode(instance);
+    if (qrcode) return qrcode;
+    await new Promise((r) => setTimeout(r, delay));
+  }
+  return null;
 };
 
 const handleOpen = async (instance) => {
   const wId = sock.user.id.split(':')[0];
+
   await connec('instances')
     .where({ instance })
     .update({
       status: 'connected',
       phone: wId,
-      qrcode_or_pairingcode: null,
+      qrcode: null,
       qr_expires_at: null,
       updated_at: connec.fn.now(),
     });
 
   console.log('✅ Conectado com sucesso:', wId);
+};
+
+const finalizeInstance = async (instance) => {
   try {
     await uploadToS3(instance);
+    await sock.logout(); // encerra a sessão com segurança
+    console.log('🔒 Sessão encerrada com sucesso após upload.');
+    await removeAuthFolder(instance); // só apaga agora
   } catch (err) {
-    console.error('❌ Erro ao mover arquivos para o S3:', err.message);
+    console.error('❌ Erro no finalizeInstance:', err.message);
   }
 };
 
-const handleClose = async (update, instance, phoneNumber) => {
+const handleClose = async (update, instance) => {
   const { lastDisconnect } = update;
   const statusCode = lastDisconnect?.error?.output?.statusCode;
   const loggedOut = statusCode === DisconnectReason.loggedOut;
+
   console.log('🔌 Conexão encerrada, por:', statusCode);
 
   if (loggedOut) {
-    const authPath = path.resolve(__dirname, '..', 'auth', instance);
-    try {
-      if (fs.existsSync(authPath)) {
-        fs.rmSync(authPath, { recursive: true, force: true });
-        console.log(`Pasta de autenticação ${authPath} removida com sucesso!`);
-      }
-    } catch (err) {
-      console.error(`❌ Erro ao remover pasta de autenticação: ${authPath}`, err.message);
-    }
+    await removeAuthFolder(instance);
   }
 
   await connec('instances')
@@ -120,25 +149,7 @@ const handleClose = async (update, instance, phoneNumber) => {
 
   if (!loggedOut) {
     console.log(`🔄 Tentando reconectar instância: ${instance}`);
-    createInstances(instance, phoneNumber);
-  }
-};
-
-const handlePairing = async (instance, phoneNumber) => {
-  try {
-    const code = await sock.requestPairingCode(`55${phoneNumber.replace(/\D/g, '')}`);
-    console.log('✅ Código de emparelhamento gerado:', code);
-
-    const payload = {
-      status: 'PairingCode',
-      qrcode_or_pairingcode: code,
-      updated_at: connec.fn.now(),
-    };
-    const exists = await connec('instances').where({ instance }).first();
-    if (exists) await connec('instances').where({ instance }).update(payload);
-    else await connec('instances').insert({ instance, ...payload });
-  } catch (err) {
-    console.error('❌ Erro ao solicitar código de emparelhamento:', err.message);
+    createInstances(instance);
   }
 };
 
@@ -146,32 +157,41 @@ const uploadToS3 = async (instance) => {
   const authPath = path.join(__dirname, '..', 'auth', instance);
 
   if (!await fs.pathExists(authPath)) {
-    console.warn(`Caminho de autenticação não encontrado: ${authPath}`);
+    console.warn(`⚠️ Caminho de autenticação não encontrado: ${authPath}`);
     return;
   }
 
-  try {
-    const files = await fs.readdir(authPath);
-    for (const file of files) {
-      const filePath = path.join(authPath, file);
-      const fileContent = await fs.readFile(filePath);
+  const files = await fs.readdir(authPath);
 
-      const command = new PutObjectCommand({
-        Bucket: process.env.AWS_S3_BUCKET,
-        Key: `auth/${instance}/${file}`,
-        Body: fileContent,
-        ContentType: 'application/json',
-      });
+  for (const file of files) {
+    const filePath = path.join(authPath, file);
+    const fileContent = await fs.readFile(filePath);
 
-      await S3.send(command);
-      console.log(`🟢 Arquivo ${file} enviado para o S3 com sucesso!`);
-    }
-  } finally {
-    await fs.remove(authPath);
-    console.log(`🧹 Pasta de autenticação ${authPath} removida com sucesso!`);
+    const command = new PutObjectCommand({
+      Bucket: process.env.AWS_S3_BUCKET,
+      Key: `auth/${instance}/${file}`,
+      Body: fileContent,
+      ContentType: 'application/json',
+    });
+
+    await S3.send(command);
+    console.log(`🟢 Arquivo ${file} enviado para o S3 com sucesso!`);
   }
-}
+};
+
+const removeAuthFolder = async (instance) => {
+  const authPath = path.resolve(__dirname, '..', 'auth', instance);
+  try {
+    if (await fs.pathExists(authPath)) {
+      await fs.remove(authPath);
+      console.log(`🧹 Pasta de autenticação ${authPath} removida com sucesso!`);
+    }
+  } catch (err) {
+    console.error(`❌ Erro ao remover pasta ${authPath}:`, err.message);
+  }
+};
 
 module.exports = {
   createInstances,
-}
+  timeForQrCode
+};
